+++ build.ts
@@ -1,0 +1,6 @@
+import { build } from 'bun';
+
+await build({
+	entrypoints: ['./src/main.ts'],
+	outdir: './dist',
+});
