+++ src/core/index.test.ts
