+++ src/systems/EntityAdapter/components/index.ts
