+++ src/systems/EntityAdapter/index.ts
