--- vite.config.ts
@@ -1,25 +1,0 @@
-// vite.config.js
-import { resolve } from 'path';
-import { defineConfig } from 'vite';
-
-export default defineConfig({
-	build: {
-		lib: {
-			// Could also be a dictionary or array of multiple entry points
-			entry: resolve(__dirname, 'lib/main.ts'),
-			name: 'SystemsCore',
-			// the proper extensions will be added
-			fileName: (format) => `systems-core.${format}.js`,
-		},
-		rollupOptions: {
-			// make sure to externalize deps that shouldn't be bundled
-			// into your library
-			external: [],
-			output: {
-				// Provide global variables to use in the UMD build
-				// for externalized deps
-				globals: {},
-			},
-		},
-	},
-});
